feat: refactor and update file structure
