+++ packages/date-picker/src/DatePicker/DatePicker.scss
@@ -3,25 +3,17 @@
 @import "~@kaizen/design-tokens/sass/spacing";
 @import "~@kaizen/design-tokens/sass/border";
 @import "~@kaizen/design-tokens/sass/shadow";
-@import "~@kaizen/draft-button/KaizenDraft/Button/styles";
-
-@import "~@kaizen/deprecated-component-library-helpers/styles/layout";
-@import "~@kaizen/deprecated-component-library-helpers/styles/color";
-@import "~@kaizen/component-library/styles/mixins/forms";
-@import "~@kaizen/draft-form/KaizenDraft/Form/Primitives/base";
 
 // Vars
 $border-radius: 7px;
 $button-height: 48px;
-$button-base-padding-horizontal: $ca-grid / 2;
+$button-base-padding-horizontal: $spacing-sm;
 $button-icon-size: 1.25rem; // 20px
 
 $disabled-opacity: 0.5;
 $placeholder-opacity: 0.7;
-$disabled-icon-opacity: 0.8;
 
-// Button styles
-.button {
+@mixin button-reset {
   appearance: none;
   display: inline;
   background: transparent;
@@ -32,27 +24,25 @@ $disabled-icon-opacity: 0.8;
   border: none;
 }
 
-.startIconAdornment {
-  margin-top: -1px;
-}
-
 .button {
-  @include form-input-reset;
-  text-align: left;
+  @include button-reset;
   position: relative;
-  background-color: $color-white;
+  display: flex;
+  align-items: center;
+  height: $button-height;
+  width: 100%;
   color: $color-purple-800;
+  font-family: $typography-paragraph-body-font-family;
+  font-size: $typography-paragraph-body-font-size;
+  font-weight: $typography-paragraph-body-font-weight;
+  line-height: 1.5;
+  letter-spacing: $typography-paragraph-body-letter-spacing;
+  text-align: left;
+  background-color: $color-white;
   background-clip: padding-box;
   border: $border-solid-border-width $border-solid-border-style $color-gray-500;
   border-radius: $border-radius;
-  display: flex;
-  font-size: 1rem;
-  line-height: 1.5;
   padding: 0 $button-base-padding-horizontal;
-  height: $button-height;
-  width: 100%;
-  margin-bottom: 2px;
-  align-items: center;
 
   &:focus:not([disabled]),
   &:hover:focus:not([disabled]),
@@ -76,8 +66,19 @@ $disabled-icon-opacity: 0.8;
   }
 
   .value {
-    @include ca-padding($start: 1.75rem, $end: $button-base-padding-horizontal);
+    padding-right: $button-base-padding-horizontal;
+    padding-left: 1.75rem;
+
+    &[dir="rtl"],
+    [dir="rtl"] & {
+      padding-right: 1.75rem;
+      padding-left: $button-base-padding-horizontal;
+    }
   }
+}
+
+.startIconAdornment {
+  margin-top: -1px;
 }
 
 // Icon adornment styles
@@ -92,14 +93,25 @@ $disabled-icon-opacity: 0.8;
 .withStartIconAdornment {
   .startIconAdornment {
     @include vertically-center-icon;
-    @include ca-position($start: $ca-grid / 2);
+    right: auto;
+    left: $spacing-sm;
+
+    &[dir="rtl"],
+    [dir="rtl"] & {
+      right: $spacing-sm;
+      left: auto;
+    }
   }
 
   .button {
-    @include ca-padding(
-      $start: $ca-grid + $button-icon-size,
-      $end: $button-base-padding-horizontal
-    );
+    padding-right: $button-base-padding-horizontal;
+    padding-left: $spacing-md + $button-icon-size;
+
+    &[dir="rtl"],
+    [dir="rtl"] & {
+      padding-right: $spacing-md + $button-icon-size;
+      padding-left: $button-base-padding-horizontal;
+    }
   }
 }
 
@@ -116,118 +128,5 @@ $disabled-icon-opacity: 0.8;
     .startIconAdornment {
       opacity: $placeholder-opacity;
     }
-  }
-}
-
-// Calendar styles
-.calendar {
-  background-color: $color-white;
-  font-family: $typography-heading-4-font-family;
-  z-index: 1;
-
-  .wrapper {
-    padding-bottom: 0;
-  }
-
-  .navbar {
-    position: absolute;
-    display: flex;
-    justify-content: space-between;
-    width: calc(100% - 3rem);
-    margin: 0 $spacing-md;
-    top: 1.25rem;
-    color: $color-purple-800;
-  }
-
-  .arrows {
-    @include button-reset;
-    display: inherit;
-    padding: 0.25rem $spacing-xs;
-    border: $border-solid-border-width $border-solid-border-style transparent;
-    border-radius: $border-radius;
-
-    &:hover {
-      cursor: pointer;
-      background-color: $color-blue-100;
-      color: $color-blue-500;
-    }
-
-    &:focus {
-      outline-offset: -2px;
-      outline: $color-blue-500 $border-focus-ring-border-style
-        $border-focus-ring-border-width;
-    }
-  }
-
-  .month {
-    margin: $spacing-md;
-    border-collapse: separate;
-    border-spacing: 1px;
-  }
-
-  .months {
-    box-shadow: $shadow-large-box-shadow;
-    border-radius: $border-solid-border-radius;
-  }
-
-  .caption {
-    text-align: center;
-    font-size: $typography-heading-4-font-size;
-    font-weight: $typography-heading-4-font-weight;
-    line-height: $typography-heading-4-line-height;
-    letter-spacing: $typography-heading-4-letter-spacing;
-    color: $color-purple-800;
-    margin: 0 0 $spacing-xs;
-  }
-
-  .weekday {
-    font-size: $typography-paragraph-body-font-size;
-    font-weight: $typography-paragraph-body-font-weight;
-    line-height: $typography-paragraph-body-line-height;
-    color: $color-purple-800;
-  }
-
-  .body {
-    font-size: $typography-paragraph-body-font-size;
-    font-weight: $typography-paragraph-body-font-weight;
-    line-height: $typography-paragraph-body-line-height;
-  }
-
-  .today.day {
-    color: $color-blue-500;
-  }
-
-  .day {
-    padding: calc(#{$spacing-md} / 3) 0.625rem;
-    border-radius: $border-radius;
-    color: $color-purple-800;
-
-    &:focus {
-      outline-offset: -2px;
-      outline: $color-blue-500 $border-focus-ring-border-style
-        $border-focus-ring-border-width;
-
-      &.disabled {
-        outline: $color-gray-300 $border-dashed-border-style
-          $border-focus-ring-border-width;
-      }
-    }
-  }
-
-  .disabled,
-  .disabled.today {
-    color: rgba($color-purple-800-rgb, 0.3);
-  }
-
-  .day:not(.interactionDisabled):not(.disabled):not(.selected):not(.outside):hover {
-    border-radius: $border-radius;
-    background-color: $color-blue-100;
-    color: $color-blue-500;
-  }
-
-  .selected:not(.disabled):not(.outside) {
-    border-radius: $border-radius;
-    background-color: $color-blue-500;
-    color: $color-white;
   }
 }
